fix: 文档目录 & source 更新bug
