+++ docs/basis-components/layer/demo/grid-layer/widgets/index.ts
