+++ docs/basis-components/layer/demo/image-layer/config/config.ts
