+++ docs/basis-components/layer/demo/image-layer/index.tsx
