+++ docs/basis-components/layer/demo/image-layer/widgets/ImageLayer.tsx
