+++ docs/basis-components/layer/demo/image-layer/widgets/index.tsx
