+++ docs/basis-components/layer/demo/layer-group/widgets/CustomLayerGroup.ts
