+++ docs/basis-components/layer/demo/point-layer/color/config/config.ts
