+++ docs/basis-components/layer/demo/point-layer/color/index.tsx
