+++ docs/basis-components/layer/demo/point-layer/color/widgets/index.tsx
