+++ docs/basis-components/layer/demo/point-layer/hover/config/config.ts
