+++ docs/basis-components/layer/demo/point-layer/hover/widgets/MyPanel.tsx
