+++ docs/basis-components/layer/demo/point-layer/hover/widgets/PointLayer.tsx
